Fix some changes
