+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".Views.Acitivity.SplashActivity"
+            android:name=".Controllers.Acitivity.SplashActivity"
             android:theme="@style/AppTheme.NoActionBar"
             >
             <intent-filter>
@@ -20,7 +20,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".Views.Acitivity.MainActivity"
+            android:name=".Controllers.Acitivity.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
