+++ app/src/main/java/com/cgwprj/parkingmanager/Controllers/Acitivity/SplashActivity.java
@@ -1,0 +1,21 @@
+package com.cgwprj.parkingmanager.Controllers.Acitivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.cgwprj.parkingmanager.R;
+
+public class SplashActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_splash);
+
+        Intent intent = new Intent(this, MainActivity.class);
+        startActivity(intent);
+        finish();
+
+    }
+}
