+++ app/src/main/java/com/cgwprj/parkingmanager/Utils/Calculator.java
@@ -17,19 +17,19 @@ public class Calculator {
         return sec_time / 60;
     }
 
-    public static int FeeCalculator(int takenTime){
+    public static int FeeCalculate(int takenTime){
 
         // Over first 10 minutes
-        if( takenTime > IntegerConstants.FEE_FREE_TIME.getConstants() ){
+        if( takenTime > FEE.FEE_FREE_TIME.getConstants() ){
 
             // First 30 minutes fee
-            int fee = IntegerConstants.FEE_FIRST_BOUNDARY_FEE.getConstants();
+            int fee = FEE.FEE_FIRST_BOUNDARY_FEE.getConstants();
 
             // Over 30 minutes.
-            if ( takenTime > IntegerConstants.FEE_FIRST_BOUNDARY_TIME.getConstants() ){
-                int overTime = takenTime - IntegerConstants.FEE_FIRST_BOUNDARY_TIME.getConstants();
+            if ( takenTime > FEE.FEE_FIRST_BOUNDARY_TIME.getConstants() ){
+                int overTime = takenTime - FEE.FEE_FIRST_BOUNDARY_TIME.getConstants();
                 int overStep = (overTime - 1)  / 10 + 1;
-                fee += ( overStep * IntegerConstants.FEE_PER_TEN_MINUTES.getConstants());
+                fee += ( overStep * FEE.FEE_PER_TEN_MINUTES.getConstants());
             }
 
             return fee;
