+++ app/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".Views.Acitivity.MainActivity"
+    tools:context=".Controllers.Acitivity.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
     <LinearLayout
