+++ app/src/main/res/layout/fragment_enroll.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Views.Fragments.EnrollFragment">
+    tools:context=".Controllers.Fragments.EnrollFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
