+++ app/src/main/res/layout/fragment_inquiry.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Views.Fragments.InquiryFragment">
+    tools:context=".Controllers.Fragments.InquiryFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
