+++ app/src/main/res/layout/fragment_lookup.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Views.Fragments.LookupFragment">
+    tools:context=".Controllers.Fragments.LookupFragment">
 
     <TextView
         android:id="@+id/lookup_text"
